pyproject.toml
[project]
name = "town-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pygltflib>=1.16.5",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
    "redis>=7.1.0",
    "pydantic>=2.12.5",
    "PyJWT>=2.11.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "aiofiles>=25.1.0",
    "zstandard>=0.22.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.26",
    "respx>=0.22",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
